src/requirements.txt
numpy==1.24.3
pandas==2.0.1
scikit-learn==1.2.2
pyod==1.0.9
azureml-defaults==1.48.0
azureml-contrib-services==1.48.0
azure-ai-ml==1.5.0
azure-identity